requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.5.0
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
easyocr==1.7.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.3
imageio==2.34.2
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.1.4
jsons==1.6.3
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
ngrok==1.3.0
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pexpect==4.8.0
piexif==1.1.3
pillow==10.3.0
psutil==6.0.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMySQL==1.1.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
python-multipart==0.0.9
pytz==2024.1
PyYAML==5.4.1
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sos==4.5.6
SQLAlchemy==2.0.31
ssh-import-id==5.11
starlette==0.37.2
subprocess.run==0.0.8
sympy==1.13.0
systemd-python==234
tifffile==2024.7.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
Twisted==22.1.0
typer==0.12.3
typing_extensions==4.12.2
typish==1.9.3
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
ultralytics==8.2.48
ultralytics-thop==2.0.0
unattended-upgrades==0.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
watchgod==0.8.2
websockets==12.0
wsproto==1.0.0
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0



# sudo apt-get update
# sudo apt-get install imagemagick

# fastapi==0.95.1
# uvicorn==0.21.1